+++ vortex-array/src/test_harness/mod.rs
@@ -14,7 +14,6 @@ use crate::arrays::bool::BoolArrayExt;
 
 #[cfg(not(codspeed))]
 pub mod trace;
-pub mod trace_arrays;
 
 /// Check that a named metadata matches its previous versioning.
 ///
